+++ src/constants/routes.ts
@@ -1,5 +1,6 @@
 const ROUTES = {
   home: "/",
+  read: "/read",
 } as const;
 
 export { ROUTES };